finalizado crud e corrigido bug de operacoes crud com id

+++ src/main/java/br/com/fiap/livraria/entity/Livro.java
@@ -14,23 +14,24 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
 import java.math.BigDecimal;
-import java.util.Date;
 
 @Getter
 @Setter
 @EqualsAndHashCode
 @AllArgsConstructor
 @NoArgsConstructor
+
 @Entity
 @Table(name = "TB_LIVRO")
 public class Livro {
+
 
     public Livro(CreateUpdateLivroDTO createUpdateLivroDTO) {
     }
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column
     private String isbn;
